Split those NAND operations into nand.c.
